Add sparse merkle tree example and add it to README.md
